requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.0.5
flask_cors
Werkzeug==3.0.1
python-dotenv==1.0.0
email-validator==2.1.0.post1
ldap3==2.9.1
openpyxl
requests